Mode marchand :  - gestion du magasin  - passage en mode marchand  - gestion des acteurs 'Merchant' TODO :  - les achats chez un marchand joueur  - gérer la reconnexion d'un marchand (unspawn + gain de kamas)

+++ trunk/Server/Stump.Server.WorldServer/Game/Actors/RolePlay/Merchants/MerchantManager.cs
@@ -1,5 +1,6 @@
 ﻿using System;
 using System.Collections.Generic;
+using System.Collections.ObjectModel;
 using System.Diagnostics.Contracts;
 using System.Linq;
 using Stump.Core.Reflection;
@@ -11,9 +12,13 @@ using MerchantSpawn = Stump.Server.WorldServer.Database.World.WorldMapMerchantRe
 
 namespace Stump.Server.WorldServer.Game.Actors.RolePlay.Merchants
 {
-    public class MerchantManager : DataManager<MerchantManager>
+    public class MerchantManager : DataManager<MerchantManager>, ISaveable
     {
         private Dictionary<int, MerchantSpawn> m_merchantSpawns;
+        private List<Merchant> m_activeMerchants = new List<Merchant>();
+        private Stack<MerchantSpawn> m_addedSpawns = new Stack<MerchantSpawn>();
+        private Stack<MerchantSpawn> m_removedSpawns = new Stack<MerchantSpawn>();
+
 
         [Initialization(InitializationPass.Sixth)]
         public override void Initialize()
@@ -26,16 +31,58 @@ namespace Stump.Server.WorldServer.Game.Actors.RolePlay.Merchants
             return m_merchantSpawns.Values.ToArray();
         }
 
-        public void AddMerchantSpawn(MerchantSpawn spawn)
+        public ReadOnlyCollection<Merchant> Merchants
         {
-            Database.Insert(spawn);
+            get { return m_activeMerchants.AsReadOnly(); }
+        }
+
+        public void AddMerchantSpawn(MerchantSpawn spawn, bool lazySave = true)
+        {
+            if (lazySave)
+                m_addedSpawns.Push(spawn);
+            else
+                Database.Insert(spawn);
             m_merchantSpawns.Add(spawn.Id, spawn);
         }
 
-        public void RemoveMerchantSpawn(MerchantSpawn spawn)
+        public void RemoveMerchantSpawn(MerchantSpawn spawn, bool lazySave = true)
         {
-            Database.Delete(spawn);
+            if (lazySave)
+                m_removedSpawns.Push(spawn);
+            else
+                Database.Delete(spawn);
             m_merchantSpawns.Remove(spawn.Id);
+        }
+
+        public void RegisterMerchant(Merchant merchant)
+        {
+            m_activeMerchants.Add(merchant);
+        }
+
+        public void UnRegisterMerchant(Merchant merchant)
+        {
+            m_activeMerchants.Remove(merchant);
+        }
+
+        public void Save()
+        {
+            while (m_addedSpawns.Count > 0)
+            {
+                var spawn = m_addedSpawns.Pop();
+                Database.Insert(spawn);
+            }
+
+            while (m_removedSpawns.Count > 0)
+            {
+                var spawn = m_removedSpawns.Pop();
+                Database.Delete(spawn);
+            }
+
+            foreach (var merchant in m_activeMerchants)
+            {
+                if (merchant.Bag.IsDirty)
+                    merchant.Save();
+            }
         }
     }
 }
